1.0.5-BETA - Rewrite of command execution because I wasn't happy with it.   - Added CommandType enum for sub commands   - Added InfractionCommand for new command handling
