+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: InfractionManager
-version: 1.0.3-BETA
+version: 1.0.5-BETA
 main: com.killerofpie.infractionmanager.InfractionManager
 
 author: KillerOfPie
